requirements.txt
pystray
pillow
mcp>=1.0.0
